pyproject.toml
[project]
name = "rpi3-fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "fastapi[standard]>=0.121.2",
    "python-dotenv>=1.0.0",
    "redis>=7.0.1",
    "uvicorn>=0.38.0",
]
